backend/pyproject.toml
[project]
name = "trading-backend"
version = "0.1.0"
description = "New FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "python-socketio>=5.11",
  "pydantic-settings>=2.2",
  "ccxt>=4.0",
  "pandas>=2.2",
  "matplotlib>=3.8",
  "mplfinance>=0.12",
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "psycopg2-binary>=2.9",
  "alembic>=1.13",
  "redis>=5.0",
  "cryptography>=42.0",
  "taskiq>=0.11",
  "taskiq-redis>=1.0",
  "httpx>=0.26",
  "certifi>=2024.2.2",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.26",
  "aiosqlite>=0.20",
]

[project.scripts]
backend-cli = "app.cli:main"

[tool.pytest.ini_options]
addopts = "-ra"
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]
